# make high precision for avg_pool and adaptive_avg_pool when data_type is float16