--- a/paddle/fluid/operators/math/pooling.h
+++ b/paddle/fluid/operators/math/pooling.h
@@ -18,6 +18,7 @@ limitations under the License. */
 
 #include "paddle/fluid/framework/eigen.h"
 #include "paddle/fluid/framework/tensor.h"
+#include "paddle/fluid/operators/amp/fp16_type_traits.h"
 #include "paddle/fluid/platform/device_context.h"
 #include "paddle/fluid/platform/hostdevice.h"
 #include "paddle/fluid/platform/macros.h"
@@ -46,29 +47,23 @@ class MaxPool {
 
 template <class T>
 class AvgPool {
-  float intermediate_res;
+  using MT = typename details::MPTypeTrait<T>::Type;
+  MT intermediate_res;
 
  public:
   DEVICE inline T initial() {
-    intermediate_res = static_cast<float>(0.0f);
+    intermediate_res = static_cast<MT>(0.0f);
     return static_cast<T>(0);
   }
-  DEVICE inline void compute(const T& x, T* y) { *y += x; }
-  DEVICE inline void finalize(const T& pool_field, T* y) { *y /= pool_field; }
+
+  DEVICE inline void compute(const T& x, T* y) {
+    intermediate_res += static_cast<MT>(x);
+  }
+
+  DEVICE inline void finalize(const T& pool_field, T* y) {
+    *y = static_cast<T>(intermediate_res / (static_cast<MT>(pool_field)));
+  }
 };
-
-template <>
-DEVICE inline void AvgPool<paddle::platform::float16>::compute(
-    const paddle::platform::float16& x, paddle::platform::float16* y) {
-  intermediate_res += static_cast<float>(x);
-}
-
-template <>
-DEVICE inline void AvgPool<paddle::platform::float16>::finalize(
-    const paddle::platform::float16& pool_field, paddle::platform::float16* y) {
-  *y = static_cast<paddle::platform::float16>(intermediate_res /
-                                              (static_cast<float>(pool_field)));
-}
 
 template <class T>
 class MaxPoolGrad {