<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
 Bug fixes
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
针对avg_pool在数据类型为float16时添加高精度设置：
前后测试结果对：
```
修复后的精度：
case0
Gradients:  [<tf.Tensor 'gradients/adaptive_average_pooling2d/split_grad/concat:0' shape=(4, 128, 128, 2048) dtype=float16>]
[paddle][adaptive_avg_pool2d] adaptive_avg_pool2d {
  run_tf: True
  run_torch: True
  repeat: 5000
  data_format: NHWC
  output_size: [2, 2]
  x_shape: [4, 128, 128, 2048]
  x_dtype: float16
  atol: 0.001
}
var pool2d_0.tmp_0 : LOD_TENSOR.shape(4, 2, 2, 2048).dtype(float16).stop_gradient(False)
Gradients:  [var x@GRAD : LOD_TENSOR.shape(4, 128, 128, 2048).dtype(float16).stop_gradient(False)]
{"name": "adaptive_avg_pool2d", "device": "GPU", "backward": true, "consistent": true, "num_outputs": 2, "diff": 0.0, "parameters": "x (Variable) - dtype: float16, shape: [4, 128, 128, 2048]\ndata_format (string): NHWC\noutput_size (list): [2, 2]\n"}

case 1

[paddle][adaptive_avg_pool2d] adaptive_avg_pool2d {
  run_tf: True
  run_torch: True
  repeat: 5000
  data_format: NHWC
  output_size: [1, 2]
  x_shape: [4, 128, 128, 2048]
  x_dtype: float16
  atol: 0.001
}
var pool2d_0.tmp_0 : LOD_TENSOR.shape(4, 1, 2, 2048).dtype(float16).stop_gradient(False)
Gradients:  [var x@GRAD : LOD_TENSOR.shape(4, 128, 128, 2048).dtype(float16).stop_gradient(False)]
{"name": "adaptive_avg_pool2d", "device": "GPU", "backward": true, "consistent": true, "num_outputs": 2, "diff": 0.0, "parameters": "x (Variable) - dtype: float16, shape: [4, 128, 128, 2048]\ndata_format (string): NHWC\noutput_size (list): [1, 2]\n"}

```
t f cas0 :
```
Gradients:  [<tf.Tensor 'gradients/adaptive_average_pooling2d/split_grad/concat:0' shape=(4, 2048, 64, 128) dtype=float32>]
[paddle][adaptive_avg_pool2d] adaptive_avg_pool2d {
  run_tf: True
  run_torch: True
  repeat: 5000
  data_format: NCHW
  output_size: [32, 32]
  x_shape: [4, 2048, 64, 128]
  x_dtype: float32
  atol: 1e-06
}
Gradients:  [var x@GRAD : LOD_TENSOR.shape(4, 2048, 64, 128).dtype(float32).stop_gradient(False)]
{"name": "adaptive_avg_pool2d", "device": "GPU", "backward": true, "consistent": true, "num_outputs": 2, "diff": 1.7881393432617188e-07, "parameters": "x (Variable) - dtype: float32, shape: [4, 2048, 64, 128]\ndata_format (string): NCHW\noutput_size (list): [32, 32]\n"}
```


tf case 1
```

Gradients:  [<tf.Tensor 'gradients/adaptive_average_pooling2d/split_grad/concat:0' shape=(4, 2048, 64, 128) dtype=float16>]
[paddle][adaptive_avg_pool2d] adaptive_avg_pool2d {
  run_tf: True
  run_torch: True
  repeat: 5000
  data_format: NCHW
  output_size: [32, 32]
  x_shape: [4, 2048, 64, 128]
  x_dtype: float16
  atol: 0.001
}
Gradients:  [var x@GRAD : LOD_TENSOR.shape(4, 2048, 64, 128).dtype(float16).stop_gradient(False)]
{"name": "adaptive_avg_pool2d", "device": "GPU", "backward": true, "consistent": true, "num_outputs": 2, "diff": 0.0, "parameters": "x (Variable) - dtype: float16, shape: [4, 2048, 64, 128]\ndata_format (string): NCHW\noutput_size (list): [32, 32]\n"}

```




base 精度
```
case 0 
Gradients:  [<tf.Tensor 'gradients/adaptive_average_pooling2d/split_grad/concat:0' shape=(4, 128, 128, 2048) dtype=float16>]
[paddle][adaptive_avg_pool2d] adaptive_avg_pool2d {
  run_tf: True
  run_torch: True
  repeat: 5000
  data_format: NHWC
  output_size: [2, 2]
  x_shape: [4, 128, 128, 2048]
  x_dtype: float16
  atol: 0.001
}
var pool2d_0.tmp_0 : LOD_TENSOR.shape(4, 2, 2, 2048).dtype(float16).stop_gradient(False)
Gradients:  [var x@GRAD : LOD_TENSOR.shape(4, 128, 128, 2048).dtype(float16).stop_gradient(False)]
---- Warning: The 0-th output (shape: (4, 2, 2, 2048), data type: float16) has diff. Detail: max_absolute_diff = 9.155e-04, max_relative_diff = 1.918e-02, offset = 28686, -0.02519 vs -0.02428, atol is 1.00e-03.



case 1
Gradients:  [<tf.Tensor 'gradients/adaptive_average_pooling2d/split_grad/concat/concat:0' shape=(4, 128, 128, 2048) dtype=float16>]
[paddle][adaptive_avg_pool2d] adaptive_avg_pool2d {
  run_tf: True
  run_torch: True
  repeat: 5000
  data_format: NHWC
  output_size: [1, 2]
  x_shape: [4, 128, 128, 2048]
  x_dtype: float16
  atol: 0.001
}
var pool2d_0.tmp_0 : LOD_TENSOR.shape(4, 1, 2, 2048).dtype(float16).stop_gradient(False)
Gradients:  [var x@GRAD : LOD_TENSOR.shape(4, 128, 128, 2048).dtype(float16).stop_gradient(False)]
---- Warning: The 0-th output (shape: (4, 1, 2, 2048), data type: float16) has diff. Detail: max_absolute_diff = 6.714e-04, max_relative_diff = 2.670e-02, offset = 14144, 0.01953 vs 0.0202, atol is 1.00e-03.

```
benchmark 失败说明：

1. 本次修改针对avg_pool 在datatype=float16类型下：将输入数据转化为float类型后进行求和运算， 相比于原始操作引入了static_cast操作，将会造成性能下降；
修改前后性能对比：

+ 使用float类型修复float16精度问题，前后性能对比，性能下降约0.5% ～ 2%；
+ 相比于pytorch性能越差20% ～  75%，对于当前网络配置中的性能与pytorch性能已经打平；
+ 对于NCHW 模型cas [4, 64, 128, 2048]e 和tensorflow性能差72.5%；
+ 当NC维度值较大，HW维度较小时，paddle性能优于pytorch

case 单位us | paddle old | paddle new | pytorch | (new-old)/old | (pytorch-new_paddle)/new_paddle
-- | -- | -- | -- | -- | --
output_size: [32, 32]  x_shape: [4, 2048, 64, 128]  x_dtype: float16 | 430.14 | 441.59 | 273.07 | 2.66% | -38.16%
output_size: [32, 32]  x_shape: [4, 128, 64, 128]  x_dtype: float16 | 31.919 | 32.738 | 24.481 | 2.57% | -25.22%
output_size: [4, 4]  x_shape: [16, 2048, 32, 32] | 119.4 | 119.41 | 387.16 | 0.01% | 224.23%
output_size: [2, 2]  x_shape: [16, 64, 112, 112]  x_dtype: float16 | 1499.6 | 1501.7 | 361.87 | 0.14% | -75.90%
output_size: [2, 2]  x_shape: [16, 256, 16, 16] | 29.835 | 29.937 | 36.68 | 0.34% | 22.52%
output_size: [2, 2]  x_shape: [16, 2048, 16, 16]  x_dtype: float16 | 61.823 | 62.085 | 211.36 | 0.42% | 240.44%
output_size[1, 1]  x_shape: [4, 64, 128, 2048] | 1111.2 | 1126.4 | 305.38 tensorflow | 1.37% | -72.52%





unsqueeze ，squeeze 性能下降与本次修改无关，通过单独测试发现并没有调用avg_pool:
```
run command: nvprof /usr/local/bin/python /paddle_test/work/benchmark/api/tests_v2/unsqueeze.py --task speed --framework paddle --json_file /paddle_test/work/benchmark/api/tests_v2/configs/unsqueeze.json --config_id 0 --check_output False --profiler none --backward True --use_gpu True --repeat 5000 --allow_adaptive_repeat False --log_level 0
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   99.47%  116.90ms     10002  11.687us  11.168us  13.408us  [CUDA memcpy DtoD]
                    0.51%  605.02us         8  75.627us  1.7280us  591.71us  [CUDA memcpy HtoD]
                    0.01%  7.1360us         4  1.7840us  1.6000us  2.3040us  [CUDA memset]
                    0.01%  6.6880us         1  6.6880us  6.6880us  6.6880us  void Eigen::internal::EigenMetaKernel<Eigen::TensorEvaluator<Eigen::TensorAssignOp<Eigen::TensorMap<Eigen::Tensor<float, int=1, int=1, long>, int=0, Eigen::MakePointer>, Eigen::TensorCwiseNullaryOp<Eigen::internal::scalar_constant_op<float>, Eigen::TensorMap<Eigen::Tensor<float, int=1, int=1, long>, int=0, Eigen::MakePointer> const > const > const , Eigen::GpuDevice>, long>(float, int=1)
```


```
squeeze.json
run command: nvprof /usr/local/bin/python /paddle_test/work/benchmark/api/tests_v2/squeeze.py --task speed --framework paddle --json_file /paddle_test/work/benchmark/api/tests_v2/configs/squeeze.json --config_id 0 --check_output False --profiler none --backward True --use_gpu True --repeat 5000 --allow_adaptive_repeat False --log_level 0
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   99.79%  12.732ms     10002  1.2720us  1.2150us  1.6960us  [CUDA memcpy DtoD]
                    0.14%  18.368us         8  2.2960us  1.7280us  5.2800us  [CUDA memcpy HtoD]
                    0.05%  6.7840us         4  1.6960us  1.6000us  1.9520us  [CUDA memset]
                    0.01%  1.6960us         1  1.6960us  1.6960us  1.6960us  void Eigen::internal::EigenMetaKernel<Eigen::TensorEvaluator<Eigen::TensorAssignOp<Eigen::TensorMap<Eigen::Tensor<float, int=1, int=1, long>, int=0, Eigen::MakePointer>, Eigen::TensorCwiseNullaryOp<Eigen::internal::scalar_constant_op<float>, Eigen::TensorMap<Eigen::Tensor<float, int=1, int=1, long>, int=0, Eigen::MakePointer> const > const > const , Eigen::GpuDevice>, long>(float, int=1)
```
2. spp, cudnn_lstm,  fake_quantize, pool_with_index 没有对应的api，所以无法在benchmark里添加测试脚本

3. avg_pool2d_1 (forward) 性能下降：本地验证：
```
 base
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   59.94%  91.027ms      5001  18.201us  17.824us  20.832us  void cudnn::detail::pooling_bw_kernel_avg<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>(cudnnTensorStruct, float const *, float const , cudnn::detail::pooling_bw_kernel_avg<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>, float const , cudnn::detail::pooling_bw_kernel_avg<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>, cudnnTensorStruct*, cudnnPoolingStruct, float, cudnnPoolingStruct, int, cudnn::reduced_divisor, float)
                   39.66%  60.229ms      5001  12.043us  9.7600us  12.896us  void cudnn::detail::pooling_fw_4d_kernel<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>(cudnnTensorStruct, float const *, cudnn::detail::pooling_fw_4d_kernel<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>, cudnnTensorStruct*, cudnnPoolingStruct, float, cudnnPoolingStruct, int, cudnn::reduced_divisor, float)

```

```
本次提交性能：
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   59.30%  90.906ms      5001  18.177us  17.856us  20.544us  void cudnn::detail::pooling_bw_kernel_avg<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>(cudnnTensorStruct, float const *, float const , cudnn::detail::pooling_bw_kernel_avg<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>, float const , cudnn::detail::pooling_bw_kernel_avg<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>, cudnnTensorStruct*, cudnnPoolingStruct, float, cudnnPoolingStruct, int, cudnn::reduced_divisor, float)
                   40.14%  61.537ms      5001  12.305us  9.6960us  12.992us  void cudnn::detail::pooling_fw_4d_kernel<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>(cudnnTensorStruct, float const *, cudnn::detail::pooling_fw_4d_kernel<float, float, cudnn::detail::averpooling_func<float>, int=2, bool=0>, cudnnTensorStruct*, cudnnPoolingStruct, float, cudnnPoolingStruct, int, cudnn::reduced_divisor, float)

```